[DS-707] Various fixes
git-svn-id: http://scm.dspace.org/svn/repo/dspace/trunk@5686 9c30dcfa-912a-0410-8fc2-9e0234be79fd

+++ dspace-api/src/main/java/org/dspace/checker/DailyReportEmailer.java
@@ -373,14 +373,11 @@ public class DailyReportEmailer
                 }
             }
 
-            if (report != null)
+            if (report != null && report.exists())
             {
-                if (report.exists())
+                if (!report.delete())
                 {
-                    if (!report.delete())
-                    {
-                        log.error("Unable to delete report file");
-                    }
+                    log.error("Unable to delete report file");
                 }
             }
         }
